Initial upgrade to RN 0.76.9. Added expo-secure-store and modules.
We have to replace react-native-sensitive-info with a new underlying mechanism as it is no longer supported nor updated for newer versions of RN. This also required me to add expo into the mix to allow us to use this. I'll write up some info in a PR on why I've chosen this route.